+++ localeurl/settings.py
@@ -2,7 +2,9 @@ import re
 from django.conf import settings
 
 SUPPORTED_LOCALES = dict(settings.LANGUAGES)
-LOCALES_RE = '|'.join(SUPPORTED_LOCALES)
+# Sort locale codes longest-first to avoid matching e.g. 'nl' before 'nl-be'
+LOCALES_RE = '|'.join(
+    sorted(SUPPORTED_LOCALES.keys(), key=lambda i: len(i), reverse=True))
 PATH_RE = re.compile(r'^/(?P<locale>%s)(?P<path>.*)$' % LOCALES_RE)
 
 LOCALE_INDEPENDENT_PATHS = getattr(settings, 'LOCALE_INDEPENDENT_PATHS', ())