+++ localeurl/tests/__init__.py
@@ -21,6 +21,7 @@ def settings_fixture(mgr):
         USE_I18N = True,
         LANGUAGES = (
             ('en', 'English'),
+            ('nl', 'Dutch'),
             ('nl-nl', 'Dutch'),
             ('nl-be', 'Flemish'),
             ('fr', 'French'),
@@ -253,7 +254,7 @@ class TagsTestCase(LocaleurlTestCase):
 
     def test_locale_url_tag(self):
         self.assertRaises(ValueError, self.render_template,
-                '{% locale_url "nl" dummy0 %}')
+                '{% locale_url "pt" dummy0 %}')
 
         self.assertEqual('/en/dummy/', self.render_template(
                 '{% locale_url "en-us" dummy0 %}'))
@@ -322,7 +323,7 @@ if localeurl_future is not None:
 
         def test_locale_url_tag(self):
             self.assertRaises(ValueError, self.render_template,
-                    '{% locale_url "nl" "dummy0" %}')
+                    '{% locale_url "pt" "dummy0" %}')
 
             self.assertEqual('/en/dummy/', self.render_template(
                     '{% locale_url "en-us" "dummy0" %}'))